refactor mocks to separate module
